requirements.txt
pytest==5.3.1
pytest-django==3.7.0
Django==2.2.8
django-model-utils==3.2.0
djangorestframework==3.10.3
redis==3.3.11
requests==2.22.0
huey==2.1.3
funcy==1.14
psycopg2==2.8.4

